Subdiffusive timescale of Rouse-GLE model.

Calculates the range of the subdiffusive timescale along with the effective subdiffusion coefficient and diffusivity constant on this range.

rouse_sub(alpha, tau, K, est_alpha = TRUE, ...)

Arguments

alpha

Rouse-GLE subdiffusion coefficient.

tau

Shortest timescale of force memory.

K

Number of modes in relaxation spectrum.

est_alpha

If FALSE, use the Rouse-GLE's alpha parameter, otherwise estimate alpha by least-squares.

...

Additional arguments to pass to prony_coeff.

Value

Vector with named elements tmin, tmax, alpha, and D.

Details

The Rouse model is a zero-mass, zero-potential GLE with force memory kernel $$ acf_F(t) = 1/K sum( exp(- t * (1:K/K)^rho / tau) ). $$ As \(K\) gets large, it exhibits has transient subdiffusion: $$ MSD(t) ~ D * t^alpha, t \in tScale t, t \notin tScale, $$ where \(alpha = 1/rho\) is the subdiffusion exponent and \(D\) is the diffusivity constant.